pyproject.toml
[project]
name = "streamoid-marketplace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2,<5.3",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "minio>=7.2.20",
    "mysqlclient>=2.2.7",
    "openpyxl>=3.1.5",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.2.1",
]

[tool.ruff]
line-length = 115

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["config", "core", "marketplace", "mapping", "seller"]
known-third-party = ["django", "rest_framework"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.11",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["streamoid"]
